Consolidated InterpolateMode for 1d and 2d
